Lab10.15 Xoa tep

+++ NODEJS-COMPLETE-GUIDE/controllers/admin.js
@@ -1,4 +1,5 @@
-const mongoose = require('mongoose');
+const mongoose = require("mongoose");
+const fileHelper = require("../util/file");
 
 const Product = require("../models/product");
 
@@ -11,7 +12,7 @@ exports.getAddProduct = (req, res, next) => {
     editing: false,
     hasError: false,
     errorMessage: null,
-    validationErrors: []
+    validationErrors: [],
   });
 };
 
@@ -32,15 +33,16 @@ exports.postAddProduct = (req, res, next) => {
         title: title,
         // imageUrl: imageUrl, //removed
         price: price,
-        description: description
+        description: description,
       },
-      errorMessage: 'Attached file is not an image',
-      validationErrors: []
+      errorMessage: "Attached file is not an image",
+      validationErrors: [],
     });
   }
 
   const errors = validationResult(req);
-  if (!errors.isEmpty()) { //catched error
+  if (!errors.isEmpty()) {
+    //catched error
     return res.status(422).render("admin/edit-product", {
       pageTitle: "Add Product",
       path: "/admin/add-product",
@@ -50,10 +52,10 @@ exports.postAddProduct = (req, res, next) => {
         title: title,
         // imageUrl: imageUrl,
         price: price,
-        description: description
+        description: description,
       },
       errorMessage: errors.array()[0].msg,
-      validationErrors: errors.array()
+      validationErrors: errors.array(),
     });
   }
 
@@ -65,7 +67,7 @@ exports.postAddProduct = (req, res, next) => {
     price: price,
     description: description,
     imageUrl: imageUrl,
-    userId: req.user
+    userId: req.user,
   });
   product
     .save()
@@ -121,7 +123,7 @@ exports.getEditProduct = (req, res, next) => {
         product: product,
         hasError: false,
         errorMessage: null,
-        validationErrors: []
+        validationErrors: [],
       });
     })
     .catch((err) => {
@@ -141,9 +143,9 @@ exports.postEditProduct = (req, res, next) => {
   const errors = validationResult(req);
 
   if (!errors.isEmpty()) {
-    return res.status(422).render('admin/edit-product', {
-      pageTitle: 'Edit Product',
-      path: '/admin/edit-product',
+    return res.status(422).render("admin/edit-product", {
+      pageTitle: "Edit Product",
+      path: "/admin/edit-product",
       editing: true,
       hasError: true,
       product: {
@@ -151,33 +153,32 @@ exports.postEditProduct = (req, res, next) => {
         // imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDescription,
-        _id: prodId
+        _id: prodId,
       },
       errorMessage: errors.array()[0].msg,
-      validationErrors: errors.array()
+      validationErrors: errors.array(),
     });
   }
 
   Product.findById(prodId)
     .then((product) => {
       if (prodId.userId.toString() !== req.user._id.toString()) {
-        return res.redirect('/');
+        return res.redirect("/");
       }
 
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDescription;
-      
-      if(image){
+
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
 
-      return product
-            .save()
-            .then((result) => {
-              console.log("Updated Product");
-              res.redirect("/admin/products");
-            }); //Ham save nay cua mongoose
+      return product.save().then((result) => {
+        console.log("Updated Product");
+        res.redirect("/admin/products");
+      }); //Ham save nay cua mongoose
     })
     .catch((err) => {
       const error = new Error(err);
@@ -187,7 +188,7 @@ exports.postEditProduct = (req, res, next) => {
 };
 
 exports.getProducts = (req, res, next) => {
-  Product.find({userId: req.user._id}) //ham find() su dung trong mongodb se khac voi find() goc trong JS. find() trong mongodb neu khong truyen dieu kien thi se tra ve toan bo du lieu
+  Product.find({ userId: req.user._id }) //ham find() su dung trong mongodb se khac voi find() goc trong JS. find() trong mongodb neu khong truyen dieu kien thi se tra ve toan bo du lieu
     // .select('title price -_id') //select trong key muon lay ra, va bo ra nhung key sau dau '-'
     // .populate('userId', 'name') // populate se lay toan bo du lieu tu collection khac co quan he theo userId truyen vao
     .then((products) => {
@@ -208,7 +209,14 @@ exports.getProducts = (req, res, next) => {
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
 
-  Product.deleteOne({_id: prodId, userId: req.user._id})//Ham findByIdAndRemove cua Mongoose
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error("Product not found!"));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id }); //Ham findByIdAndRemove cua Mongoose
+    })
     .then(() => {
       console.log("Destroyed Product");
       res.redirect("/admin/products");
